【免测】将 mip-sandbox 的沙盒环境生成封装成一个函数 (#259)
* 修复不同组件库构建出来的组件样式覆盖的问题

* 增加沙盒构造函数，让 mip 核心决定何时构造沙盒
