+++ packages/mip-sandbox/package.json
@@ -1,6 +1,6 @@
 {
   "name": "mip-sandbox",
-  "version": "1.0.22",
+  "version": "1.0.23",
   "description": "sandbox tools for MIP project",
   "main": "lib/sandbox.js",
   "scripts": {
